Move chain_id from Tx to Block on mock  (#439)
* update: Move `chain_id` from Tx to Block on mock

When generating a `TestContext` now, users will be able to set the
`chain_id` parameter only at block level but not at tx one.

Anyway, `MockTransaction` will still have this field. But there's no
public API to interact with it. Instead, the values of the `chain_id`
for the txs of a block will automatically be set to the value that was
set as `chain_id` for the block.

This closes the circle on the turn arround that was done in
https://github.com/appliedzkp/zkevm-specs/pull/168 by adding `CHAINID`
as a blockCtx Opcode rather than a Tx one for simplification &
optimization purposes.

Resolves: #437

* update: Update Cargo.lock

* fix: Reference MockTransaction instead of MockBlock