Box ExecutionConfig in EvmCircuit (#583)
The type ExecutionConfig has been growing a lot with each opcode
gadget that we add.  Currently it's around 337 KiB.  By having functions
that return an EvmCircuit or a struct that contains an EvmCircuit,
we're actually allocating 337 KiB in the stack to hold the
ExecutionConfig.

So if we have a TestCircuit::configure that retruns a struct containing
EvmCircuit and calls EvmCircuit::configure that returns a struct
containing ExecutionConfig and calls ExecutionConfig::configure that
returns itself, by this point we have 3 times the size of
ExecutionConfig allocated in the stack, which means 1 MiB.

By moving this struct to the heap with Box as soon as possible, we will
only have one call in the entire call stack that requires allocating the
struct in the stack.

It may be interesting to look for other big structs that we hold in the
stack in several calls within a call stack.